build: Switch to native ARM64 compilation and fix Docker/build issues
- Switch CI/Release linux-arm64 builds to use ubuntu-24.04-arm runner for native compilation.
- Remove cross-compilation setup (gcc-aarch64-linux-gnu, multi-arch deps) in favor of native gcc.
- Update Dockerfile for compatibility (debian:bookworm, V source install, mkdir /app/bin).
- Update CI/Release workflows to use ubuntu-latest for amd64.
- Cleanup unused variables and imports in V source code.

+++ src/infra/compression/lz4.v
@@ -207,7 +207,7 @@ fn lz4_decompress_block(compressed_data []u8) ![]u8 {
 				return error('incomplete match')
 			}
 
-			offset := int(compressed_data[ip]) | ((int(b) & 0xe0) << 3)
+			offset := int(compressed_data[ip]) | (int(u32(b) & 0xe0) << 3)
 			ip++
 
 			mut match_len := int(compressed_data[ip])
